backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
python-dotenv
# LangChain core
langchain

# Cohere integration
langchain-cohere
langchain-community
